Move out layout context to allow HMR
